requirements.txt
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
matplotlib>=3.7.0
numpy>=1.24.0
Pillow>=9.5.0
pyyaml>=6.0
tqdm>=4.65.0
huggingface_hub>=0.16.0
datasets>=2.14.0